Move shipment resolvers into dedicated modules
